# Conflicts:
#	accounts/tests.py
#	accounts/views.py
#	requirements.txt
#	sounds/views.py

+++ forum/views.py
@@ -25,14 +25,13 @@ from django.contrib.auth.decorators import login_required, permission_required
 from django.contrib.auth.models import User
 from django.http import HttpResponseRedirect, Http404, \
     HttpResponsePermanentRedirect
-from django.shortcuts import render, get_object_or_404
+from django.shortcuts import render, get_object_or_404, redirect
 from django.template import loader
 from django.urls import reverse
 from django.db import transaction
 
 from forum.forms import PostReplyForm, NewThreadForm, PostModerationForm
 from forum.models import Forum, Thread, Post, Subscription
-from utils.functional import combine_dicts
 from utils.mail import send_mail_template
 from utils.pagination import paginate
 from utils.search.search_forum import add_post_to_solr, delete_post_from_solr
@@ -88,7 +87,8 @@ class last_action(object):
 @last_action
 def forums(request):
     forums = Forum.objects.select_related('last_post', 'last_post__author', 'last_post__thread').all()
-    return render(request, 'forum/index.html', locals())
+    tvars = {'forums': forums}
+    return render(request, 'forum/index.html', tvars)
 
 
 @last_action
@@ -98,10 +98,12 @@ def forum(request, forum_name_slug):
     except Forum.DoesNotExist:
         raise Http404
 
+    tvars = {'forum': forum}
     paginator = paginate(request, Thread.objects.filter(forum=forum, first_post__moderation_state="OK")
                          .select_related('last_post', 'last_post__author'), settings.FORUM_THREADS_PER_PAGE)
+    tvars.update(paginator)
 
-    return render(request, 'forum/threads.html', combine_dicts(locals(), paginator))
+    return render(request, 'forum/threads.html', tvars)
 
 
 @last_action
@@ -127,7 +129,11 @@ def thread(request, forum_name_slug, thread_id):
         except Subscription.DoesNotExist:
             pass
 
-    return render(request, 'forum/thread.html', combine_dicts(locals(), paginator))
+    tvars = {'thread': thread,
+             'forum': forum,
+             'has_subscription': has_subscription}
+    tvars.update(paginator)
+    return render(request, 'forum/thread.html', tvars)
 
 
 @last_action
@@ -136,7 +142,9 @@ def latest_posts(request):
                          Post.objects.select_related('author', 'author__profile', 'thread', 'thread__forum')
                          .filter(moderation_state="OK").order_by('-created').all(), settings.FORUM_POSTS_PER_PAGE)
     hide_search = True
-    return render(request, 'forum/latest_posts.html', combine_dicts(locals(), paginator))
+    tvars = {'hide_search': hide_search}
+    tvars.update(paginator)
+    return render(request, 'forum/latest_posts.html', tvars)
 
 
 @last_action
@@ -227,7 +235,11 @@ def reply(request, forum_name_slug, thread_id, post_id=None):
         messages.add_message(request, messages.INFO, "You're not allowed to post in the forums because your account "
                                                      "has been temporaly blocked after multiple spam reports")
 
-    return render(request, 'forum/reply.html', locals())
+    tvars = {'forum': forum,
+             'thread': thread,
+             'form': form,
+             'latest_posts': latest_posts}
+    return render(request, 'forum/reply.html', tvars)
 
 
 @login_required
@@ -255,7 +267,7 @@ def new_thread(request, forum_name_slug):
                     add_post_to_solr(post)
                     set_to_moderation = False
 
-                # Add first post to thread (first post will allways be the same)
+                # Add first post to thread (first post will always be the same)
                 # We need to reload thread object from DB, not so overwrite the object we created before when saving
                 updated_thread = Thread.objects.get(id=thread.id)
                 updated_thread.first_post = post
@@ -280,7 +292,9 @@ def new_thread(request, forum_name_slug):
         messages.add_message(request, messages.INFO, "You're not allowed to post in the forums because your account "
                                                      "has been temporaly blocked after multiple spam reports")
 
-    return render(request, 'forum/new_thread.html', locals())
+    tvars = {'forum': forum,
+             'form': form}
+    return render(request, 'forum/new_thread.html', tvars)
 
 
 @login_required
@@ -329,7 +343,8 @@ def old_topic_link_redirect(request):
 def post_delete(request, post_id):
     post = get_object_or_404(Post, id=post_id)
     if post.author == request.user or request.user.has_perm('forum.delete_post'):
-        return render(request, 'forum/confirm_deletion.html', locals())
+        tvars = {'post': post}
+        return render(request, 'forum/confirm_deletion.html', tvars)
     else:
         raise Http404
 
@@ -337,16 +352,22 @@ def post_delete(request, post_id):
 @login_required
 def post_delete_confirm(request, post_id):
     post = get_object_or_404(Post, id=post_id)
-    if post.author == request.user or request.user.has_perm('forum.delete_post'):
-        thread = post.thread
-        post.delete()
-        try:
-            return HttpResponseRedirect(
-                reverse('forums-post', args=[thread.forum.name_slug, thread.id, thread.last_post.id]))
-        except (Post.DoesNotExist, Thread.DoesNotExist, AttributeError) as e:
-            return HttpResponseRedirect(reverse('forums-forums'))
-    else:
-        raise Http404
+    if request.method == 'POST':
+        if post.author == request.user or request.user.has_perm('forum.delete_post'):
+            thread = post.thread
+            forum = thread.forum
+            post.delete()
+            # If the post was the only post in the thread, redirect to the forum
+            try:
+                thread.refresh_from_db()
+            except Thread.DoesNotExist:
+                return redirect('forums-forum', forum_name_slug=forum.name_slug)
+            try:
+                return redirect('forums-post', thread.forum.name_slug, thread.id, thread.last_post.id)
+            except (Post.DoesNotExist, Thread.DoesNotExist, AttributeError):
+                return HttpResponseRedirect(reverse('forums-forums'))
+
+    raise Http404
 
 
 @login_required
@@ -365,7 +386,8 @@ def post_edit(request, post_id):
                     reverse('forums-post', args=[post.thread.forum.name_slug, post.thread.id, post.id]))
         else:
             form = PostReplyForm(request, '', {'body': post.body})
-        return render(request, 'forum/post_edit.html', locals())
+        tvars = {'form': form}
+        return render(request, 'forum/post_edit.html', tvars)
     else:
         raise Http404
 
@@ -388,7 +410,6 @@ def moderate_posts(request):
                     messages.add_message(request, messages.INFO, 'The user has been successfully deleted.')
                 except User.DoesNotExist:
                     messages.add_message(request, messages.INFO, 'The user has already been deleted.')
-                    pass
             elif action == "Delete Post":
                 post.delete()
                 messages.add_message(request, messages.INFO, 'The post has been successfully deleted.')
@@ -397,6 +418,8 @@ def moderate_posts(request):
     post_list = []
     for p in pending_posts:
         f = PostModerationForm(initial={'action': 'Approve', 'post': p.id})
-        post_list.append({'post': p,'form': f})
-    forums = True  # prevent base template showing forum search
-    return render(request, 'forum/moderate.html', locals())
+        post_list.append({'post': p, 'form': f})
+
+    tvars = {'post_list': post_list,
+             'hide_search': True}
+    return render(request, 'forum/moderate.html', tvars)
